requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==4.0.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
gcloud==0.18.3
googleapis-common-protos==1.54.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jws==0.1.3
oauth2client==4.1.3
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
pyparsing==3.0.7
python-jwt==2.0.1
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
